services/speech-ai/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1
pydantic-settings==2.3.0
httpx==0.27.0
redis==5.0.7
python-jose[cryptography]==3.3.0
prometheus-fastapi-instrumentator==7.0.0
opentelemetry-sdk==1.25.0
opentelemetry-instrumentation-fastapi==0.46b0
structlog==24.2.0
python-multipart==0.0.9
